Get Philadelphia District Attorney's Office Charges Data

This function grabs the publicly available from the Philadelphia District Attorney's Office Public Data Dashboard.

get_charges(
  geography = c("citywide", "police_district", "zip_code", "census_tract")
)

Arguments

geography
A string with one of the following: 'citywide', 'police_district', 'zip_code', 'census_tract'.

Value

A data.frame with the data at the desired geographic unit.

Examples

data <- get_charges("citywide")
head(data)

#>         date homicide rape robbery_gun robbery_other aggravated_assault_gun
#> 1 2014-01-01        0    1           1             0                      0
#> 2 2014-01-02        0    3          14            12                     11
#> 3 2014-01-03        0    1           8             2                      1
#> 4 2014-01-04        0    0          22             1                      7
#> 5 2014-01-05        0    0           1             2                      1
#> 6 2014-01-06        1    0           0             7                      7
#>   aggravated_assault_other other_assaults sexual_assault_and_other_sex_offenses
#> 1                       17             12                                     1
#> 2                       11             15                                     0
#> 3                        7             13                                     0
#> 4                        6              2                                     0
#> 5                        5              7                                     2
#> 6                        9             10                                     0
#>   burglary_residential burglary_commercial theft_of_motor_vehicle_tag
#> 1                    5                   0                          0
#> 2                    1                   0                          0
#> 3                    0                   1                          0
#> 4                    1                   0                          0
#> 5                    0                   0                          0
#> 6                    1                   1                          0
#>   theft_from_person theft_from_auto retail_theft theft auto_theft
#> 1                 0               1            4     1          3
#> 2                 0               2            5     0          3
#> 3                 0               0            8     0          2
#> 4                 0               4            4     1          2
#> 5                 0               1            3     0          0
#> 6                 0               0            3     3          5
#>   fraud_theft_of_services embezzlement drug_possession drug_sales dui
#> 1                       1            0               8          1  22
#> 2                       3            0               3          8   9
#> 3                       1            0               9         14   7
#> 4                       1            0              12         11   8
#> 5                       0            1               5          1   6
#> 6                       4            0               7          6   7
#>   possession_of_firearms prostitution_sex_work patronizing_prostitutes
#> 1                      7                     0                       0
#> 2                      0                     0                       0
#> 3                      0                     0                       0
#> 4                      1                     0                       0
#> 5                      6                     0                       0
#> 6                      3                     0                       0
#>   threats_of_violence uncategorized_offenses
#> 1                   0                      9
#> 2                   2                      9
#> 3                   0                      8
#> 4                   2                     12
#> 5                   0                      1
#> 6                   1                      7